+++ src/main/java/io/github/kamitejp/platform/linux/wlroots/dependencies/slurp/SlurpFade.java
@@ -1,0 +1,6 @@
+package io.github.kamitejp.platform.linux.wlroots.dependencies.slurp;
+
+public enum SlurpFade {
+  FADE,
+  NO_FADE;
+}