+++ src/main/java/io/github/kamitejp/platform/linux/wlroots/dependencies/wlrctl/WlrctlResult.java
@@ -1,0 +1,10 @@
+package io.github.kamitejp.platform.linux.wlroots.dependencies.wlrctl;
+
+public sealed interface WlrctlResult
+  permits WlrctlResult.ExecutionFailed,
+          WlrctlResult.Error,
+          WlrctlResult.Ok {
+  record ExecutionFailed() implements WlrctlResult {}
+  record Error(String error) implements WlrctlResult {}
+  record Ok() implements WlrctlResult {}
+}